Move preload files into dedicated `src/preload` directory
