+++ src/preload/api/helpers.ts
