+++ src/preload/api/menu.ts
