+++ src/preload/api/misc.ts
